+++ libatalk/cnid/db3/cnid_db3_update.c
@@ -1,5 +1,5 @@
 /*
- * $Id: cnid_db3_update.c,v 1.1.4.2.2.1 2005/01/30 20:56:22 didg Exp $
+ * $Id: cnid_db3_update.c,v 1.1.4.2.2.2 2005/09/27 10:40:41 didg Exp $
  */
 
 #ifdef HAVE_CONFIG_H
@@ -30,7 +30,7 @@
  * handle the did/name data, there are a bunch of functions to get
  * and set the various fields. */
 int cnid_db3_update(struct _cnid_db *cdb, const cnid_t id, const struct stat *st,
-                const cnid_t did, char *name, const int len
+                const cnid_t did, char *name, const size_t len
                 /*, const char *info, const int infolen*/)
 {
     CNID_private *db;
